requirements.txt
cdsapi
jupyter
matplotlib
netCDF4
os
pandas
python-dotenv
rasterio